requirements.txt
Django>=5.0,<6.0
djangorestframework
requests
psycopg2-binary
gunicorn
